feat: add self-deployed Multicall3 contract
